+++ Assets/OxGFrame/AssetLoader/Scripts/Runtime/Bundle/YooAssets/RemoteServices.cs
@@ -1,0 +1,27 @@
+﻿using System.IO;
+using YooAsset;
+
+namespace OxGFrame.AssetLoader.Bundle
+{
+    public class HostServers : IRemoteServices
+    {
+        private string _hostServer;
+        private string _fallbackHostServer;
+
+        public HostServers(string hostServer, string fallbackHostServer)
+        {
+            this._hostServer = hostServer;
+            this._fallbackHostServer = fallbackHostServer;
+        }
+
+        public string GetRemoteMainURL(string fileName)
+        {
+            return Path.Combine(this._hostServer, fileName);
+        }
+
+        public string GetRemoteFallbackURL(string fileName)
+        {
+            return Path.Combine(this._fallbackHostServer, fileName);
+        }
+    }
+}